Chronos Update: Merge SIMS code with latest Chronos
1. Added ETLUtil component to the whole project (this is similar to
previous when SIMS team add this component to Chronos-2.2.0)
2. Rewrote and added the logic of hostname check into MesosJobFramewokr
3. Add “config” field as part of definition of BaseJob class, and add
this field to JobSerializer/JobDeserializer, and MesosTaskBuilder to
support “config”
4. Add a editable web UI section for Job config.
5. Update JobScheduler, so that, a job's config will be merged with its
prerequisite job’s config.
6. Update MesosTaskBuilder for executor. This part is tricky. For
executors, if I directly use the logic from published code, it does not
work (Chronos will not be able to schedule jobs). Previous developers
changed this part to hard-code “/usr/bin/etl_script.sh” for executor to
make the code work, and I am following this logic. My option is, still
checking in the working one (with hard-coded string) to SIMS teams
Chronos repository, but if we want to RI to Chronos online (airbnb
repository), we need to get rid of this hard-coded string.
7. Fix a bug in previous code (in ETLUtils.java), use origConfig
(instead of empty string) as initial value when override or merge
config. So that if JSON parsing throws an error, we still keep the
original config, instead of overwrite the original config with empty
string.